♻ Refactorisation Commune
